The beginning starts with some light
These are the files for the initial version of the advent-of-code
template. The idea behind this is to provide tooling for people
willing to start the Advent of Code challenge.

This is intended to provide an easy way to setup the basic files for
the puzzles in the challenge and, using the advent-of-code-data
package, run it against the adventofcode.com API so the results can
easily be validated and submitted.